requirements.txt
colorama==0.4.6
pyfiglet==0.8.post1
requests==2.31.0
python-whois==0.8.0
dnspython==2.4.2
beautifulsoup4==4.12.2
shodan==1.30.1
python-nmap==0.7.1
tqdm==4.66.1
cryptography==41.0.5
Flask==3.0.0
Flask-WTF==1.2.1
Werkzeug==3.0.1
selenium==4.15.2
webdriver-manager==4.0.1
urllib3==2.1.0
lxml==4.9.3
python-dotenv==1.0.0
PyPDF2==3.0.1
reportlab==4.1.0
waybackpy==3.0.6
aiohttp==3.9.3
asyncio==3.4.3
pyOpenSSL==23.3.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
PySocks==1.7.1
aiohappyeyeballs>=2.3.0
aiosignal>=1.1.2
attrs>=17.3.0
frozenlist>=1.1.1
multidict<7.0,>=4.5
propcache>=0.2.0
yarl<2.0,>=1.17.0
websocket-client==1.6.4
PyJWT==2.8.0
paramiko==3.3.1
scapy==2.5.0
pycryptodome==3.19.0
pyyaml==6.0.1
pyasn1==0.5.1
bcrypt==4.0.1
cffi==1.16.0
pycparser==2.21
pynacl==1.5.0
six==1.16.0
typing-extensions==4.8.0
websockets==12.0
cryptography==41.0.5
pyOpenSSL==23.3.0
dnspython==2.4.2
requests[socks]==2.31.0
python-whois==0.8.0
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3==2.1.0
certifi==2023.11.17